+++ lib/cgraph/write.c
@@ -254,22 +254,6 @@ char *agcanonStr(char *str)
     return agstrcanon(str, buffer);
 }
 
-/**
- * canonicalize a string for printing.
- * If html is true, use HTML canonicalization.
- * Shared static buffer - unsafe.
- */
-char *agcanon(char *str, int html)
-{
-    char* buf = getoutputbuffer(str);
-    if (buf == NULL)
-	return NULL;
-    if (html)
-	return agcanonhtmlstr(str, buf);
-    else
-	return _agstrcanon(str, buf);
-}
-
 static int _write_canonstr(Agraph_t *g, iochan_t *ofile, char *str, bool chk) {
     if (chk) {
 	str = agcanonStr(str);